refactor: del Box component
