--- src/components/showMore/Box/styled.ts
@@ -1,12 +1,0 @@
-import { Animated } from 'react-native';
-
-import styled from '@emotion/native';
-
-export const BoxContainer = styled.View``;
-
-export const BoxWrapper = styled(Animated.View)`
-  border-radius: 16px;
-  width: 100%;
-  margin-top: 4px;
-  /* border: 1px solid red; */
-`;
