+++ src/components/showMore/index.ts
@@ -1,1 +1,2 @@
-export * from './Box';
+export * from './Section';
+export * from './SectionItem';
